deepen: fold StalenessDetector into the StalenessBanner module
Replaces the shallow lib/validation/staleness-detector.ts module (73 lines, a
two-method static class) with private constants inside its sole production
caller, components/ui/email/templates/sections/StalenessBanner.tsx. The
banner now computes daysOld and severity inline; the StalenessDetector seam
is gone.

Why this was shallow:

  - StalenessDetector exposed `check()` and `formatRelativeTime()` static
    methods. The banner only ever read result.isStale + result.daysOld +
    result.severity from check(); result.message was never rendered (the
    banner inlines its own message string directly in JSX) and
    formatRelativeTime() was used ONLY internally by check() and by tests.
  - Production callers: exactly one — StalenessBanner.
  - Test file __tests__/email/staleness-detection.test.ts asserted on
    StalenessDetector.check()'s result fields and on
    StalenessDetector.formatRelativeTime() directly. Those tests targeted
    the shallow seam, not the rendered banner.
  - One adapter, one caller — by the "one adapter means a hypothetical seam"
    rule (process/Deepening/deepening.md), the seam wasn't load-bearing.

What changed:

  - Inlined the 7-day STALE_THRESHOLD_DAYS and 30-day CRITICAL_THRESHOLD_DAYS
    constants and the daysOld math directly into the banner component.
  - Removed the unused result.message and formatRelativeTime() — the banner
    has always rendered its own message string. Removing them shrinks the
    surface from a 4-field result type + a relative-time formatter (used
    nowhere except tests) to a single integer comparison.
  - Deleted lib/validation/staleness-detector.ts.
  - Deleted __tests__/email/staleness-detection.test.ts.
  - Moved the "StalenessBanner position (above EmailLeadHeader)" describe
    block from the deleted test file into staleness-banner.test.tsx — it
    was already a rendered-JSX interface test, so it belongs alongside the
    banner's other interface tests.
  - Added six rendered-JSX interface tests to staleness-banner.test.tsx
    covering the threshold (fresh / 5d / 8d / 18d / 30d / 73d) and the
    warning→critical severity escalation (asserting on the border hex,
    which innerHTML retains — per CLAUDE.md #10 background/color hex are
    converted to rgb() and unreliable).

Dependency category: in-process (per process/Deepening/deepening.md §1).
Pure date-math code merged into its only caller. No adapter strategy needed.

Tests deleted: 1 file (__tests__/email/staleness-detection.test.ts, 119 lines
— the StalenessDetector seam tests).

Tests added/relocated at the new interface (staleness-banner.test.tsx,
8 cases total, all passing):

  StalenessBanner
    ✓ does not render when filing is fresh (same day)
    ✓ does not render when filing is 5 days old (below 7-day threshold)
    ✓ renders banner for 8-day-old filing
    ✓ renders warning-severity border for 18-day-old filing (amber)
    ✓ escalates to critical-severity border at 30+ days (red)
    ✓ flips to critical exactly at the 30-day boundary
  StalenessBanner position (above EmailLeadHeader)
    ✓ renders stale banner before the logo + headline for an 8-day-old filing
    ✓ does not render the banner for a fresh filing (< 7 days old)

Full __tests__/email run: 63 passed, 0 failed (992 tests).

ADRs respected: ADR-0002 (analysis-depth inlining) — same anti-pattern, same
fix shape. No new ADR written; the CONTEXT.md "Staleness Banner" entry
covers the design.

CONTEXT.md updated: added a Staleness Banner entry alongside the existing
deep-module entries, documenting the severity thresholds, the inline-message
contract, and the positional regression captured in the banner tests.
