+++ CONTEXT.md
@@ -54,6 +54,9 @@ The post-parse redaction that scans every prose field of a [Summary] (`whyItMatt
 ### Filing Prompt
 The system+user+JSON-schema bundle handed to the LLM to summarize one [Filing]. Implemented as one deep module (`lib/ai/prompts/unified-prompts.ts`) behind two functions — `generateFilingPrompt(formType, content, options)` and `getSchemaForFormType(formType)` — plus the `FORM_SCHEMAS` registry. The module dispatches form-type-specific schema and journalist-tone instructions internally; callers (`lib/ai/summarize.ts`, `lib/ai/filing-analyzer.ts`, `lib/ai/parsers/simple-parser.ts`, `lib/ai/summarization/chunk-processor.ts`) never branch on form type. Replaces the legacy dual-prompt system — a `PromptTemplate` base class plus five `Form10KPrompt` / `Form10QPrompt` / `Form8KPrompt` / `FormForm4Prompt` / `GenericFilingPrompt` subclasses (each ~100 LOC of inlined system+user+schema strings) — which had drifted from production for months: zero non-test callers, and the journalist-tone copy lived in two places (the subclasses AND `unified-prompts.ts`) with the unified copy as the only one production actually used. The subclasses' tests asserted on internal substrings of `getSystemPrompt()` output — past the interface, not behaviour — so deleting them removed test mass that wasn't anchored to anything the system shipped.
 
+### Staleness Banner
+The amber/red email banner that warns recipients when a [Summary] was delivered significantly after the [Filing]'s original SEC filing date. Renders nothing for filings <7 days old; warning (amber) for 7–29 days; critical (red) at 30+ days. Implemented as one deep module (`components/ui/email/templates/sections/StalenessBanner.tsx`) behind a single `<StalenessBanner filingDate now? />` JSX interface — the daysOld diff, severity threshold, and the inline message string live as private constants inside the component. Replaces a former `lib/validation/staleness-detector.ts` `StalenessDetector` class with `check()` + `formatRelativeTime()` static methods that the banner was the sole production caller of; the split exposed a shallow seam (one adapter, one caller) whose tests asserted on the seam's `result.message` / `result.daysOld` / `result.severity` fields rather than on observable rendered output. The deepened module is verified through `__tests__/email/staleness-banner.test.tsx` against rendered JSX (text plus border-hex retention, per CLAUDE.md #10), including the positional regression test that locks in the banner-above-EmailLeadHeader order across the minimalist templates (10-K, 10-Q, 8-K, Form 4). Same shape as [Why It Matters Guard] / [Historical Context] / ADR-0002.
+
 ### Tier Scheduling
 How a [Subscription] tier drives the cron pipeline, along two axes that intentionally classify the same tier strings differently. The **priority axis** sets where a job sits in the queue once enqueued (MAX→9, PRO→7, FREE/unknown→5; active trials borrow MAX's 9), with a per-form-type materiality bonus folded in by `getCompositePriority` (8-K most urgent). The **cadence axis** decides whether a user is due for processing and how often (`normalizeTier` buckets both MAX and PRO into one processing frequency — 5 min — while FREE falls to HOBBY's 120 min). Implemented as one deep module (`lib/cron/tier-eligibility.ts`) behind the priority/cadence function set; trial eligibility is delegated to `lib/auth/tier-eligibility.ts` (`isActiveTrial`), the cross-cutting authority shared with auth and summarization. Absorbs the former `tier-priority.ts`, so a maintainer changing tier behaviour reads one place and can see that the two axes disagree on MAX by design.
