+++ components/ui/email/templates/sections/StalenessBanner.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { StalenessDetector } from '../../../../../lib/validation/staleness-detector';
 
 interface StalenessBannerProps {
   filingDate: Date;
@@ -8,22 +7,35 @@ interface StalenessBannerProps {
 }
 
 /**
- * Email banner component that warns users when a filing summary
- * was delivered significantly after the original filing date.
+ * Email banner that warns users when a filing summary was delivered
+ * significantly after the original filing date.
  *
- * Renders nothing for fresh filings (< 7 days old).
- * Shows an amber warning banner for stale filings.
+ * Renders `null` for fresh filings (< 7 days old). For 7–29 days renders the
+ * amber warning banner; at 30+ days the banner escalates to red (critical).
+ *
+ * The staleness rule is intentionally inline rather than behind a separate
+ * StalenessDetector seam — there's exactly one production caller (this
+ * banner, embedded in the minimalist templates) and the prior
+ * extraction-for-testability split (formerly `lib/validation/staleness-detector.ts`)
+ * exposed a shallow interface that tests targeted directly. Coverage now
+ * flows through this component's JSX output via
+ * `__tests__/email/staleness-banner.test.tsx`. Same shape as ADR-0002.
  */
-export function StalenessBanner({ filingDate, now }: StalenessBannerProps) {
-  const result = StalenessDetector.check(filingDate, now);
+const STALE_THRESHOLD_DAYS = 7;
+const CRITICAL_THRESHOLD_DAYS = 30;
+const MS_PER_DAY = 24 * 60 * 60 * 1000;
 
-  if (!result.isStale) {
+export function StalenessBanner({ filingDate, now = new Date() }: StalenessBannerProps) {
+  const daysOld = Math.floor((now.getTime() - filingDate.getTime()) / MS_PER_DAY);
+
+  if (daysOld < STALE_THRESHOLD_DAYS) {
     return null;
   }
 
-  const bgColor = result.severity === 'critical' ? '#FEE2E2' : '#FEF3C7';
-  const borderColor = result.severity === 'critical' ? '#EF4444' : '#F59E0B';
-  const textColor = result.severity === 'critical' ? '#991B1B' : '#92400E';
+  const isCritical = daysOld >= CRITICAL_THRESHOLD_DAYS;
+  const bgColor = isCritical ? '#FEE2E2' : '#FEF3C7';
+  const borderColor = isCritical ? '#EF4444' : '#F59E0B';
+  const textColor = isCritical ? '#991B1B' : '#92400E';
 
   return (
     <table width="100%" cellPadding="0" cellSpacing="0" style={{ marginBottom: '12px' }}>
@@ -38,7 +50,7 @@ export function StalenessBanner({ filingDate, now }: StalenessBannerProps) {
             lineHeight: '1.4',
             color: textColor,
           }}>
-            This summary was delayed — the filing was originally filed {result.daysOld} days ago.
+            This summary was delayed — the filing was originally filed {daysOld} days ago.
           </td>
         </tr>
       </tbody>
